Change nphase evaluation

+++ src/ffiasm/fft.hpp
@@ -42,13 +42,17 @@ public:
     void fft(Element *a, u_int64_t n);
     void fft2(Element *a, u_int64_t n);
     void ifft(Element *a, u_int64_t n);
+
+    // block implementations
     void fft_block(Element *dst, Element *src, u_int64_t n, u_int64_t ncols, u_int64_t nphase = NUM_PHASES_, u_int64_t nblock = NUM_BLOCKS_);
     void ifft_block(Element *dst, Element *src, u_int64_t n, u_int64_t ncols, u_int64_t nphase = NUM_PHASES_, u_int64_t nblock = NUM_BLOCKS_);
-    template <int N_, int NCOLS_, int NBLOCK_, int NCOLS_BLOCK_, int NPHASE_>
+
+    // template functions
+    template <uint64_t N_, uint64_t NCOLS_, uint64_t NBLOCK_, uint64_t NCOLS_BLOCK_, uint64_t NPHASE_>
     void fft_block_divisible(Element *dst, Element *src);
-    template <int N_, int NCOLS_BLOCK_, int NCOLS_, int NPHASE_>
+    template <uint64_t N_, uint64_t NCOLS_BLOCK_, uint64_t NCOLS_, uint64_t NPHASE_>
     void fft_block_iters_divisible(Element *dst, Element *src, u_int64_t offset_cols, Element *aux);
-    template <int N_, int NCOLS_BLOCK_, int NCOLS_>
+    template <uint64_t N_, uint64_t NCOLS_BLOCK_, uint64_t NCOLS_>
     void reversePermutation_block_divisible(Element *dst, Element *src, u_int64_t offset_cols);
 
     u_int32_t log2(u_int64_t n);
